+++ docs/docs/release_notes.md
@@ -2,6 +2,7 @@
 - Add all() query
 - Add tests for `pantherdb`, `load_configs()`, `status.py`, `Panel`, `multipart`, `request headers`
 - Refactor `Headers()` class
+- Check `uvloop` installation on `Panther init`
 - Minor Improvement
 
 ### 3.2.1